Make the part 2 hints a bit more vague.
Now that we're having the students do Rosalind over the summer, I think
we want the rama project to be a bit more difficult.  I thought a good
way to do this was to not break down part 2 so much, because hopefully
that will force the students to break down the problem on their own a
bit more, and that's an important skill to practice.